fix(web): review fixes — extract nav-utils, untrack $effect, migrate $app/state
- Extract isActive(), labelForSlug(), buildBreadcrumbs() to nav-utils.ts
- Fix $effect reactivity bug: wrap body in untrack() so only pathname
  triggers mobile sidebar close (prevents open→immediate close loop)
- Migrate $app/stores → $app/state (rune-native, SvelteKit 2.12+)
- Settings tab: use startsWith for defensive nested sub-route matching
